requirements.txt
-r ai_agent/requirements.txt
networkx==3.4.2
qutip==5.0.4
watchdog==6.0.0
firebase-admin==6.5.0

# Server requirements
fastapi[standard]==0.115.12
aiofiles==24.1.0
python-dotenv==1.1.0
uvicorn==0.34.2
redis-om==0.3.5
celery==5.5.3
firebase-admin==6.5.0
watchdog